Return after throwing an error
When we get into a situation where Nan::ThrowError() needs to be called,
there is no coming back. Albeit, that doesn't stop the execution of the
function it is called from. Moreover, it swallows the error. This commit
fixes this behaviour by making sure the function explicitly returns
after throwing the error.